Fix bug on closing when using DX input without a joystick attached. Run Memory Vis on nes frames instead of with timer. Add Sound Visualizer. Change square freq calculation. Simulate reads and writes after NSFs return instead of just empty cycles for DMC fetch.

+++ DirectXEmu/EmuoTron/Channels/Noise.cs
@@ -127,11 +127,11 @@ namespace EmuoTron.Channels
                     volume = envelope;
                 else
                     volume = envelopeCounter;
-                byte feedback;
+                int feedback;
                 if (loop)
-                    feedback = (byte)((shiftReg & 1) ^ ((shiftReg >> 6) & 1));
+                    feedback = (shiftReg & 1) ^ ((shiftReg >> 6) & 1);
                 else
-                    feedback = (byte)((shiftReg & 1) ^ ((shiftReg >> 1) & 1));
+                    feedback = (shiftReg & 1) ^ ((shiftReg >> 1) & 1);
                 shiftReg >>= 1;
                 shiftReg = (ushort)(shiftReg | (feedback << 14));
                 divider = timer;
